Recriada validação de hora e corrigido bug
Corrigido bug quando dia estava marcado e não havia o campo de hora in e
out.

+++ app/controllers/UserController.php
@@ -46,20 +46,20 @@ class UserController extends \BaseController {
 				'email' => 'required|email|unique:users',
 				'password' => 'required|confirmed|min:6',
 				'level' => 'required|integer|level_check',
-				'day_0_time_in' => 'TimeFormat',
-				'day_0_time_out' => 'TimeFormat',
-				'day_1_time_in' => 'TimeFormat',
-				'day_1_time_out' => 'TimeFormat',
-				'day_2_time_in' => 'TimeFormat',
-				'day_2_time_out' => 'TimeFormat',
-				'day_3_time_in' => 'TimeFormat',
-				'day_3_time_out' => 'TimeFormat',
-				'day_4_time_in' => 'TimeFormat',
-				'day_4_time_out' => 'TimeFormat',
-				'day_5_time_in' => 'TimeFormat',
-				'day_5_time_out' => 'TimeFormat',
-				'day_6_time_in' => 'TimeFormat',
-				'day_6_time_out' => 'TimeFormat'
+				'day_0_time_in' => 'sometimes|required|TimeFormat',
+				'day_0_time_out' => 'sometimes|required|TimeFormat',
+				'day_1_time_in' => 'sometimes|required|TimeFormat',
+				'day_1_time_out' => 'sometimes|required|TimeFormat',
+				'day_2_time_in' => 'sometimes|required|TimeFormat',
+				'day_2_time_out' => 'sometimes|required|TimeFormat',
+				'day_3_time_in' => 'sometimes|required|TimeFormat',
+				'day_3_time_out' => 'sometimes|required|TimeFormat',
+				'day_4_time_in' => 'sometimes|required|TimeFormat',
+				'day_4_time_out' => 'sometimes|required|TimeFormat',
+				'day_5_time_in' => 'sometimes|required|TimeFormat',
+				'day_5_time_out' => 'sometimes|required|TimeFormat',
+				'day_6_time_in' => 'sometimes|required|TimeFormat',
+				'day_6_time_out' => 'sometimes|required|TimeFormat'
 			)
 		);
 
@@ -130,31 +130,6 @@ class UserController extends \BaseController {
 	 */
 	public function update($id)
 	{
-		$validator = Validator::make(
-			Input::all(), 
-			Array(
-				'day_0_time_in' => 'timeFormat',
-				'day_0_time_out' => 'timeFormat',
-				'day_1_time_in' => 'timeFormat',
-				'day_1_time_out' => 'timeFormat',
-				'day_2_time_in' => 'timeFormat',
-				'day_2_time_out' => 'timeFormat',
-				'day_3_time_in' => 'timeFormat',
-				'day_3_time_out' => 'timeFormat',
-				'day_4_time_in' => 'timeFormat',
-				'day_4_time_out' => 'timeFormat',
-				'day_5_time_in' => 'timeFormat',
-				'day_5_time_out' => 'timeFormat',
-				'day_6_time_in' => 'timeFormat',
-				'day_6_time_out' => 'timeFormat'
-			)
-		);
-
-
-		if ($validator->fails())
-		{
-			return Redirect::back()->withInput()->withErrors($validator);
-		}
 		$user = User::findOrFail($id);
 		$user_time = UsersTimes::where('user_id',$id);
 
@@ -164,7 +139,21 @@ class UserController extends \BaseController {
 				'name' => 'required',
 				'surname' => 'required',
 				'password' => 'confirmed|min:6',
-				'level' => 'required|integer|level_check'
+				'level' => 'required|integer|level_check',
+				'day_0_time_in' => 'sometimes|required|TimeFormat',
+				'day_0_time_out' => 'sometimes|required|TimeFormat',
+				'day_1_time_in' => 'sometimes|required|TimeFormat',
+				'day_1_time_out' => 'sometimes|required|TimeFormat',
+				'day_2_time_in' => 'sometimes|required|TimeFormat',
+				'day_2_time_out' => 'sometimes|required|TimeFormat',
+				'day_3_time_in' => 'sometimes|required|TimeFormat',
+				'day_3_time_out' => 'sometimes|required|TimeFormat',
+				'day_4_time_in' => 'sometimes|required|TimeFormat',
+				'day_4_time_out' => 'sometimes|required|TimeFormat',
+				'day_5_time_in' => 'sometimes|required|TimeFormat',
+				'day_5_time_out' => 'sometimes|required|TimeFormat',
+				'day_6_time_in' => 'sometimes|required|TimeFormat',
+				'day_6_time_out' => 'sometimes|required|TimeFormat'
 			)
 		);
 
@@ -179,7 +168,7 @@ class UserController extends \BaseController {
 
 
 		for($n=0;$n<7;$n++){
-			if (Input::get("day_check_$n") == 1) {
+			if ( (Input::get("day_check_$n") == 1) && (Input::has(array("day_${n}_time_in", "day_${n}_time_out"))) ) {
 				UsersTimes::where('user_id', $id)->where('weekday', $n)->delete();
 
 				$weekday = new UsersTimes();
